+++ R/PooledConnectionHandler.R
@@ -28,6 +28,7 @@ requiredPackage <- function(packageName) {
   }
 }
 
+
 # map a ConnectionDetails objects and translates them to DBI pool args
 .DBCToDBIArgs <- list(
   "sqlite" = function(cd) {
@@ -35,20 +36,6 @@ requiredPackage <- function(packageName) {
     list(
       drv = RSQLite::SQLite(),
       dbname = cd$server()
-    )
-  },
-  "postgresql" = function(cd) {
-    requiredPackage("RPostgres")
-    host <- strsplit(cd$server(), "/")[[1]][1]
-    dbname <- strsplit(cd$server(), "/")[[1]][2]
-    list(
-      drv = RPostgres::Postgres(),
-      dbname = dbname,
-      host = host,
-      user = cd$user(),
-      port = cd$port(),
-      password = cd$password(),
-      options = "sslmode=require"
     )
   },
   "duckdb" = function(cd) {
@@ -78,13 +65,12 @@ requiredPackage <- function(packageName) {
   }
 )
 
-
 #' Pooled Connection Handler
 #'
 #' @description
 #' Transparently works the same way as a standard connection handler but stores pooled connections.
 #' Useful for long running applications that serve multiple concurrent requests.
-#'
+#' Note that a side effect of using this is that each call to this increments the .GlobalEnv attribute `RMMPooledHandlerCount`
 #' @importFrom pool dbPool poolClose
 #' @importFrom DBI dbIsValid
 #' @importFrom withr defer
@@ -94,7 +80,15 @@ PooledConnectionHandler <- R6::R6Class(
   classname = "PooledConnectionHandler",
   inherit = ConnectionHandler,
   private = list(
-    dbConnectArgs = NULL
+    dbConnectArgs = NULL,
+    .handlerId = 1,
+    activeConnectionRefs = list(),
+    .returnPooledObject = function(frame) {
+      if (!is.null(attr(frame, self$getCheckedOutConnectionPath(), exact = TRUE))) {
+        pool::poolReturn(attr(frame, self$getCheckedOutConnectionPath(), exact = TRUE))
+        attr(frame, self$getCheckedOutConnectionPath()) <- NULL
+      }
+    }
   ),
   public = list(
     #' @param connectionDetails             DatabaseConnector::connectionDetails class
@@ -126,10 +120,14 @@ PooledConnectionHandler <- R6::R6Class(
       private$dbConnectArgs <- dbConnectArgs
       self$connectionDetails <- connectionDetails
       self$snakeCaseToCamelCase <- snakeCaseToCamelCase
+      handlerCount <- attr(.GlobalEnv, "RMMPooledHandlerCount", exact = TRUE)
+      private$.handlerId <- ifelse(is.null(handlerCount), 1, handlerCount + 1)
+
       if (loadConnection) {
         self$initConnection()
       }
     },
+
     #' initialize pooled db connection
     #' @description
     #' Overrides ConnectionHandler Call
@@ -144,43 +142,141 @@ PooledConnectionHandler <- R6::R6Class(
       self$isActive <- TRUE
     },
 
+    #' Used for getting a checked out connection from a given environment (if one exists)
+    #' @param .deferedFrame  defaults to the parent frame of the calling block.
+    getCheckedOutConnectionPath = function() {
+      return(paste0("RMMcheckedOutConnection", private$.handlerId))
+    },
+
     #' Get Connection
     #' @description
     #' Returns a connection from the pool
     #' When the desired frame exits, the connection will be returned to the pool
+    #' As a side effect, the connection is stored as an attribute within the calling frame (e.g. the same function) to prevent multiple
+    #' connections being spawned, which limits performance.
+    #'
+    #' If you call this somewhere you need to think about returning the object or you may create a connection that
+    #' is never returned to the pool.
     #' @param .deferedFrame  defaults to the parent frame of the calling block.
     getConnection = function(.deferedFrame = parent.frame(n = 2)) {
-      conn <- pool::poolCheckout(super$getConnection())
-      withr::defer(pool::poolReturn(conn), envir = .deferedFrame)
-      return(conn)
+      checkmate::assertEnvironment(.deferedFrame)
+      if (!self$dbIsValid()) {
+        self$initConnection()
+      }
+
+      # Equivalent to pool::localCheckout but uses only a single connection in tha calling frame
+      if (is.null(attr(.deferedFrame, self$getCheckedOutConnectionPath(), exact = TRUE))) {
+        attr(.deferedFrame, self$getCheckedOutConnectionPath()) <- pool::poolCheckout(pool = self$con)
+
+        # Store reference to active frame
+        private$activeConnectionRefs[[length(private$activeConnectionRefs) + 1]] <- .deferedFrame
+        withr::defer(
+          {
+            private$.returnPooledObject(.deferedFrame)
+          },
+          envir = .deferedFrame
+        )
+      }
+
+      return(attr(.deferedFrame, self$getCheckedOutConnectionPath(), exact = TRUE))
     },
 
     #' get dbms
     #' @description Get the dbms type of the connection
     dbms = function() {
-      conn <- self$getConnection()
-      DatabaseConnector::dbms(conn)
+      return(self$connectionDetails$dbms)
     },
 
     #' Close Connection
     #' @description
-    #' Overrides ConnectionHandler Call
+    #' Overrides ConnectionHandler Call - closes all active connections called with getConnection
     closeConnection = function() {
       if (self$dbIsValid()) {
+        # Return any still active pooled objects
+        lapply(private$activeConnectionRefs, private$.returnPooledObject)
         pool::poolClose(pool = self$con)
       }
       self$isActive <- FALSE
       self$con <- NULL
     },
 
+
+    #' queryDb
+    #' @description
+    #' query database and return the resulting data.frame
+    #'
+    #' If environment variable LIMIT_ROW_COUNT is set Returned rows are limited to this value (no default)
+    #' Limit row count is intended for web applications that may cause a denial of service if they consume too many
+    #' resources.
+    #' @param sql                                   sql query string
+    #' @param snakeCaseToCamelCase                  (Optional) Boolean. return the results columns in camel case (default)
+    #' @param overrideRowLimit                      (Optional) Boolean. In some cases, where row limit is enforced on the system
+    #'                                              You may wish to ignore it.
+    #' @param ...                                   Additional query parameters
+    #' @returns boolean TRUE if connection is valid
+    queryDb = function(sql, snakeCaseToCamelCase = self$snakeCaseToCamelCase, overrideRowLimit = FALSE, ...) {
+      # Limit row count is intended for web applications that may cause a denial of service if they consume too many
+      # resources.
+      if (!self$isActive) {
+        self$initConnection()
+      }
+
+      sql <- .limitRowCount(sql, overrideRowLimit)
+      sql <- self$renderTranslateSql(sql, ...)
+      conn <- pool::poolCheckout(self$con)
+      on.exit(pool::poolReturn(conn))
+      tryCatch(
+        {
+          data <- self$queryFunction(sql, snakeCaseToCamelCase = snakeCaseToCamelCase, connection = conn)
+        },
+        error = function(error) {
+          if (self$dbms() %in% c("postgresql", "redshift")) {
+            self$executeFunction("ABORT;", conn)
+          }
+          stop(paste0(sql, "\n\n", error))
+        }
+      )
+
+      return(data)
+    },
+
+    #' executeSql
+    #' @description
+    #' execute set of database queries
+    #' @param sql                                   sql query string
+    #' @param ...                                   Additional query parameters
+    executeSql = function(sql, ...) {
+      if (!self$isActive) {
+        self$initConnection()
+      }
+
+      sql <- self$renderTranslateSql(sql, ...)
+      conn <- pool::poolCheckout(self$con)
+      on.exit(pool::poolReturn(conn))
+      tryCatch(
+        {
+          self$executeFunction(sql, conn)
+        },
+        error = function(error) {
+          if (self$dbms() %in% c("postgresql", "redshift")) {
+            self$executeFunction("ABORT;", conn)
+          }
+          stop(paste0(sql, "\n\n", error))
+        }
+      )
+
+      return(invisible(NULL))
+    },
+
+
     #' query Function
     #' @description
     #' Overrides ConnectionHandler Call. Does not translate or render sql.
     #' @param sql                                   sql query string
+    #' @param connection                                   db connection assumes pooling is handled outside of call
     #' @param snakeCaseToCamelCase                  (Optional) Boolean. return the results columns in camel case (default)
-    queryFunction = function(sql, snakeCaseToCamelCase = self$snakeCaseToCamelCase) {
-      conn <- self$getConnection()
-      data <- DatabaseConnector::dbGetQuery(conn, sql, translate = FALSE)
+    queryFunction = function(sql, snakeCaseToCamelCase = self$snakeCaseToCamelCase, connection) {
+      data <- DatabaseConnector::dbGetQuery(connection, sql, translate = FALSE)
       if (snakeCaseToCamelCase) {
         colnames(data) <- SqlRender::snakeCaseToCamelCase(colnames(data))
       } else {
@@ -193,9 +289,9 @@ PooledConnectionHandler <- R6::R6Class(
     #' @description
     #' Overrides ConnectionHandler Call. Does not translate or render sql.
     #' @param sql                                   sql query string
-    executeFunction = function(sql) {
-      conn <- self$getConnection()
-      DatabaseConnector::dbExecute(conn, sql, translate = FALSE)
+    #' @param connection                                  DatabaseConnector connection. Assumes pooling is handled outside of call
+    executeFunction = function(sql, connection) {
+      DatabaseConnector::dbExecute(connection, sql, translate = FALSE)
     }
   )
 )