+++ tests/testthat/test-PooledConnection.R
@@ -19,9 +19,22 @@ test_that("internal connection handlers", {
   conn <- do.call(pool::dbPool, args)
   expect_class(conn, "Pool")
   pool::poolClose(pool = conn)
+})
 
-  if (Sys.getenv("CDM5_POSTGRESQL_SERVER") != "") {
-    args <- .DBCToDBIArgs$postgresql(testDatabaseConnectionDetails)
-    expect_list(args)
-  }
+test_that("Maintained connection consistency", {
+  testthat::expect_null(attr(parent.frame(n = 1), "RMMcheckedOutConnection1"))
+  pch <- PooledConnectionHandler$new(testDatabaseConnectionDetails)
+  conn <- pch$getConnection()
+  # This test is a proxy for when the frame exits the pooled object will be returned
+  # Note, this is not guaranteed
+  testthat::expect_identical(conn, attr(parent.frame(n = 1), "RMMcheckedOutConnection1"))
+})
+
+test_that("Abort Connections", {
+  skip_if_results_db_not_available()
+  pch <- PooledConnectionHandler$new(testDatabaseConnectionDetails)
+  expect_error(pch$queryDb("BROKEN"))
+  expect_equal(pch$queryDb("SELECT 1 as c")$c, 1)
+  expect_error(pch$executeSql("MORE BROKEN"))
+  pch$executeSql("CREATE TABLE #@rnd (id int); DROP TABLE #@rnd;", rnd = testSchema)
 })